requirements.txt
crewai
langchain
langchain-openai
python-dotenv
pyyaml